requirements.txt
interception_python==0.0.1
keyboard==0.13.5
Pymem==1.10.0
slider==0.7.0
